+++ packages/liveblocks-core/src/ai.ts
@@ -949,6 +949,7 @@ export type Ai = {
       name: string,
       chatId?: string
     ): DerivedSignal<AiOpaqueToolDefinition | undefined>;
+    statusΣ: Signal<Status>;
   };
   /** @private This API will change, and is not considered stable. DO NOT RELY on it. */
   getChatById: (chatId: string) => AiChat | undefined;
@@ -1012,6 +1013,8 @@ export function createAi(config: AiConfig): Ai {
     knowledgeStore,
   };
 
+  const statusΣ = new Signal<Status>("initial");
+
   // Delta batch processing system to throttle incoming delta updates. Incoming
   // deltas are buffered and only let through every every 25ms. This creates
   // a ceiling of max 40 rerenders/second during streaming.
@@ -1046,7 +1049,7 @@ export function createAi(config: AiConfig): Ai {
   }
 
   let lastTokenKey: string | undefined;
-  function onStatusDidChange(_newStatus: Status) {
+  function onStatusDidChange(newStatus: Status) {
     const authValue = managedSocket.authValue;
     if (authValue !== null) {
       const tokenKey = getBearerTokenFromAuthValue(authValue);
@@ -1069,6 +1072,9 @@ export function createAi(config: AiConfig): Ai {
         }
       }
     }
+
+    // Forward to the outside world
+    statusΣ.set(newStatus);
   }
   let _connectionLossTimerId: TimeoutID | undefined;
   let _hasLostConnection = false;
@@ -1439,6 +1445,7 @@ export function createAi(config: AiConfig): Ai {
         getChatMessagesForBranchΣ:
           context.messagesStore.getChatMessagesForBranchΣ,
         getToolΣ: context.toolsStore.getToolΣ,
+        statusΣ,
       },
 
       getChatById: context.chatsStore.getChatById,