+++ packages/liveblocks-core/src/index.ts
@@ -303,7 +303,6 @@ export type {
   BroadcastedEventServerMsg,
   CommentsEventServerMsg,
   InitialDocumentStateServerMsg,
-  RejectedStorageOpServerMsg,
   RoomStateServerMsg,
   ServerMsg,
   UpdatePresenceServerMsg,